Do not autoload RSpec formatter and remove monkey patch
We don't want to force the BDD formatter just because the gem is loaded
in the Gemfile because developers might want to use different
formatters.

Also, the BDD formatter was implemented by monkey patching and
overriding the behavior of RSpec Documenation Formatter, which is also
not desired since we want to add functionality instead of override.